Add files via upload
In this project, understanding how waiting time affects people and how frequently they must wait for a cab during the booking process. We also want to observe the behaviour of drivers and encourage customers to pay in cash instead of online. The focus of our analysis is on reducing waiting time while prioritizing the safety and security of our customers.